requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-ckeditor==6.1.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==3.2
Markdown==3.3.4
Pillow==8.3.2
psycopg2-binary
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
pydantic
channels[daphne]
channels_redis
django_redis
sentry-sdk
pyTelegramBotAPI
